perf: Phase 1 — array transposition table (bounded memory, ~30% faster)
Replace the std::unordered_map transposition table with a fixed-size
256MB flat array (index = hash & mask, depth-preferred eviction on
collision; each slot stores the full key to detect collisions). The
old map grew without bound and OOM'd (Aborted()) on billion-node
searches — a 12-empty all-different position now COMPLETES instead of
crashing. Eviction is sound: a miss just recomputes the same value, and
the bound-flag fix (29ac9ae) means move order no longer leaks into scores.

Also widen the node counter int -> long long: it wrapped past INT_MAX on
billion-node searches (a 12-empty showed -1.36B nodes). Cosmetic — never
affected the score — but the editor's node readout was garbage on 12+.

Measured (Node): 11-empty 8.5s vs ~12s, 5.0M nps vs ~3.25M (~30% faster);
12-empty completes (~2.93B nodes, ~420s) with no OOM.

Gate: PASS (correctness vs frozen baseline + determinism + consistency
on all stress positions). Root hexuki.wasm is byte-identical to the
gate-verified binary.

+++ c++engine/src/ai/mcts.cpp
@@ -244,7 +244,7 @@ double MCTS::simulate(HexukiBitboard& board, const MCTSConfig& config) {
                 // Cache builds up across simulations → later sims are much faster
                 int searchDepth = emptyHexes;
                 int currentPlayer = board.getCurrentPlayer();
-                int nodesSearched = 0;
+                long long nodesSearched = 0;
 
                 // Create local killer/history tables for this simulation
                 minimax::KillerMoves killers;
